### 2017-04-21 #246 @arfon
👋 @noamross - would this be a suitable submission for ROpenSci?
