### 2017-04-21 #246 @noamross
Thanks @arfon. I think this is out of scope for us, as it's a a generic data munging tool: https://github.com/ropensci/onboarding/blob/master/policies.md
