### 2017-04-21 #246 @arfon
> Thanks @arfon. I think this is out of scope for us, as it's a a generic data munging tool: https://github.com/ropensci/onboarding/blob/master/policies.md

OK thanks @noamross.

@karthik - you were suggested by the submitting author as an editor for this submission. Is that OK with you?
