### 2017-04-24 2f1e583
Improve paper.md.
