### 2017-04-24 38b79ff
Improvement to paper.md.
