+++ paper/paper.md
@@ -22,6 +22,6 @@ bibliography: paper.bib
 
 # Summary
 
-This is an R package to fill in some functionality that I wish were provided in either 'base', 'stringr' or 'stringi'. For example, R has no 'file.move', just 'file.copy' and 'file.rename' so the best way to move a file was to cleverly rename it. 'filesstrings' provides the easier 'MoveFiles'. Perhaps the best feature of the package is its ability to extract numbers from strings. Making further use of this functionality, it also has a convenient function for fixing file names such that their numbering is consistent with alphabetical order. For example, 'file10.txt' comes before 'file9.txt' in alphabetical order, so the 'NiceFileNums' function would recognise this and rename them to 'file10.txt' and 'file09.txt' respectively. See https://github.com/rorynolan/filesstrings or https://cran.r-project.org/package=filesstrings.
+This is an R package to fill in some functionality that is not provided in either 'base', 'stringr' or 'stringi'. For example, R has no 'file.move', just 'file.copy' and 'file.rename' so the best way to move a file was to cleverly rename it. 'filesstrings' provides the easier 'MoveFiles'. Perhaps the best feature of the package is its ability to extract numbers from strings. Making further use of this functionality, it also has a convenient function for fixing file names such that their numbering is consistent with alphabetical order. For example, 'file10.txt' comes before 'file9.txt' in alphabetical order, so the 'NiceFileNums' function recognises this and renames them to 'file10.txt' and 'file09.txt' respectively. See https://github.com/rorynolan/filesstrings or https://cran.r-project.org/package=filesstrings.
 
 # References
