### 2017-05-13 #246 @rorynolan
Hello again,
My reviewer suggestions are @jsta, @masalmon, @benmarwick and @cdcrabtree.
In reality, any avid R user would be appropriate, the package is quite general and easy to understand.
Rory
