### 2017-05-15 #246 @benmarwick
Review of filesstrings

I'm using the [ropensci reviewer template](https://github.com/ropensci/onboarding/blob/master/reviewer_template.md) here:

- [x] As the reviewer I confirm that there are no conflicts of interest for me to review this work

#### Documentation

y: yes
i: could be improved
n: no

- [i] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
- [i] **Installation instructions:** for the development version of package and any non-standard dependencies in README
- [i] **Vignette(s)** demonstrating major functionality that runs successfully locally
- [y] **Function Documentation:** for all exported functions in R help
- [y] **Examples** for all exported functions in R Help that run successfully locally
- [i] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and `URL`, `Maintainer` and `BugReports` fields in DESCRIPTION

>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [i] **A short summary** describing the high-level functionality of the software
>- [y] **Authors:**  A list of authors with their affiliations
>- [i] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [n] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

#### Functionality

- [y] **Installation:** Installation succeeds as documented.
- [y] **Functionality:** Any functional claims of the software been confirmed.
- [y] **Performance:** Any performance claims of the software been confirmed.
- [y] **Automated tests:** Unit tests cover essential functions of the package
- [y] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines
- [y] **Software license**:There is an OSI approved license included in the repository. (this is not part of the ropensci reviewer checklist, but I've added it for completeness)

#### Final approval (post-review)

Estimated hours spent reviewing: 3

---

### Review Comments

This is an expertly-written R package that adheres to many conventions of good practice and efficient programming. It contributes some useful functions in R for file, directory, and string manipulation that could be very useful for many users.

The statement of need should be more specific, I recommend something like: "This package provides convenient functions for moving files, deleting directories, removing spaces from file names, and a variety of string operations that I wished were easier in base R."

Installation instructions should show how to install dev version, e.g. "To install the development version , run:
devtools::install_github("rorynolan/filesstrings")"

Vignettes are present. The main issue with the vignettes is that they are not very useful, especially the Files vignette. The heading in both vignettes, "Please Read the Manual", is a bit puzzling - the reader is already reading the manual (i.e. the vignette), so why ask them to do this? The examples in the Files vignette are too artificial, and don't represent a typical user's experience. I recommend to consult vignettes some widely used packages and re-write to match their style and accessibility for the typical user. I've edited the readme to indicate text that is more user-friendly.

Recommend to add contribution guidelines in the README or a CONTRIBUTING file, with a link from the readme

In paper.md and elsewhere, 'cleverly' doesn't seem like the right way to describe file moving in base R. 'Unintuitive' seems more appropriate. There are no references, which is unusual. Recommend to take a look at R packages that have been accepted by JOSS to see what they put as references.

I will make a pull request to help with some of these suggestions.

Now I'm using the [ropensci pkg guidelines](https://github.com/ropensci/packaging_guide/blob/master/README.md) here:

## Package naming

filesstrings with the ss in the middle seems a bit awkward. Why not filestrings?

## Function/variable naming & general syntax

ropensci recommends `snake_case`, and the base functions that this pkg complements use period.separated. But in this pkg we have UpperCamelCase, which is unexpected, and will likely make it harder for the user to recall the function names (it's one of the [least common naming conventions](https://journal.r-project.org/archive/2012-2/RJournal_2012-2_Baaaath.pdf))

It seems more natural and easier to remember if `MoveFiles` is named `move.files`. Similarly, `RemoveDirs` should be `dir.delete` to echo `dir.create`.

It seems more nature for `PutFilesInDir` to be a default behaviour of `MoveFiles` when the directory name doesn't exit

##  README

✔ The readme is good, I've made some suggestions to make the text more user-friendly

##  Code of Conduct

Recommend to add, cf. the ropensci guidelines:

* We recommend that you use a code of conduct such as the [Contributor Covenant](http://contributor-covenant.org/) in developing your package.  You can document your code of conduct in a `CODE_OF_CONDUCT.md` or `CONDUCT.md` file in the package root directory, and linking to this file from the `README.md` file.  `devtools::use_code_of_conduct()` will add the Contributor Covenant template to your package.

## Documentation

✔ Uses roxygen2

##  News

✔ There is a NEWS.md

Recommend to use sections as needed, including: `NEW FEATURES`, `MINOR IMPROVEMENTS`,
`BUG FIXES`, `DEPRECATED AND DEFUNCT`. Under each header list items as needed. For each item give
a description of the new feature, improvement, bug fix, or deprecated function/feature. Link
to any related GitHub issue like `(#12)`. The `(#12)` will resolve on GitHub in Releases to a
link to that issue in the repo.

## Authorship

✔ The `DESCRIPTION` file does list the package authors and contributors to a package, using the `Authors@R` syntax to indicate their roles (author/creator/contributor etc.)

## Testing

✔✔ There is an extensive set of tests, using testthat
## Versioning

✔ Uses semantic versioning.

✔ Uses Git tags

##  Continuous integration

✔ Uses continuous integration.

## Examples

✔ Includes extensive examples in the documentation.
## Package dependencies

✔ Uses `Imports` instead of `Depends`

## Console messages

Recommend to add some message()s to confirm to the user that a function completed successfully. For example, something like "37 files moved, no errors"
