### 2017-05-15 #260 @whedon
[REVIEW]: filesstrings: An R package for file and string manipulation
**Submitting author:** @rorynolan (<a href="http://orcid.org/0000-0002-5239-4043">Rory Nolan</a>)
**Repository:** <a href="https://github.com/rorynolan/filesstrings" target ="_blank">https://github.com/rorynolan/filesstrings</a>
**Version:** v0.4.1
