### 2017-05-15 #246 @karthik
@benmarwick Thanks so much! At JOSS we review on a separate thread. Could you move your comments over to #260, please?
