### 2017-05-15 3b124f3
remove dplyr dependency
