### 2017-05-17 #260 @rorynolan
Dear Ben,
Thank you for taking the time to review my package. I agree with all of your points and will amend my package accordingly. I'll have it ready in a few weeks.
Regarding the renaming of the package, since it's already on CRAN, I think I'll leave the name as it is, even though filestrings would be better for sure.
Regarding the snake_case naming of functions, I made the package according to google's R style guide. I now use snake_case in everything I do, but when I switched styles I was knee-deep in this package. It already has over a thousand downloads on CRAN, so do you think it would do more good than harm for me to restyle it? If you think it's a good idea, I'm willing to.
Thanks again,
Rory
