### 2017-05-18 dc5338d
Make changes suggested in JOSS review.
