+++ man/ExtractNumbers.Rd
@@ -13,10 +13,10 @@ ExtractNumbers(string, leave.as.string = FALSE, decimals = FALSE,
 ExtractNonNumerics(string, decimals = FALSE, leading.decimals = FALSE,
   negs = FALSE)
 
-NthNumber(string, n, leave.as.string = FALSE, decimals = FALSE,
+NthNumber(string, n = 1, leave.as.string = FALSE, decimals = FALSE,
   leading.decimals = FALSE, negs = FALSE)
 
-NthNonNumeric(string, n, decimals = FALSE, leading.decimals = FALSE,
+NthNonNumeric(string, n = 1, decimals = FALSE, leading.decimals = FALSE,
   negs = FALSE)
 }
 \arguments{
@@ -35,8 +35,8 @@ the start of a number?}
 are not handled here (see the examples).}
 
 \item{n}{The index of the number (or non-numeric) that you seek. Negative
-indexing is allowed i.e. \code{n = 1} will give you the first number (or
-non-numeric) whereas \code{n = -1} will give you the last number (or
+indexing is allowed i.e. \code{n = 1} (the default) will give you the first
+number (or non-numeric) whereas \code{n = -1} will give you the last number (or
 non-numeric), \code{n = -2} will give you the second last number and so on.}
 }
 \value{
