### 2017-05-18 e86b126
Remove dplyr from strings vignette.
