### 2017-05-18 1e5310c
Delete SplitCamelcase.Rd
