+++ man/SplitCamelcase.Rd
@@ -1,27 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/strings.R
-\name{SplitCamelCase}
-\alias{SplitCamelCase}
-\title{Split a string based on camelcase}
-\usage{
-SplitCamelCase(string, lower = FALSE)
-}
-\arguments{
-\item{string}{A character vector.}
-
-\item{lower}{Do you want the output to be all lower case (or as is)?}
-}
-\value{
-A list of character vectors, one list element for each element of
-\code{string}.
-}
-\description{
-Vectorised over \code{string}.
-}
-\examples{
-SplitCamelCase(c("RoryNolan", "NaomiFlagg", "DepartmentOfSillyHats"))
-}
-\references{
-Adapted from Ramnath Vaidyanathan's answer at
-http://stackoverflow.com/questions/8406974/splitting-camelcase-in-r.
-}
