### 2017-05-18 #260 @rorynolan
Hi @benmarwick
Apart from renaming the package and changing the style, I've implemented everything you said and pushed the changes to github.
In particular, I've
* tried to improve the vignettes
* structured the NEWS file as suggested
* added function names `file.move()` and `dir.delete()` to comply with base R naming style
* changed the paper and README to include a better statement of purpose
* added instructions to README to include how to install dev version
* made `PutFilesInDir()` to be a default behaviour of `MoveFiles()`
* added references to the paper (my bibtex references weren't integrating properly into paper.md so I put them in explicitly)
* added a CONTRIBUTING file and mentioned it in my README
* added the informative `message`s
Let me know what you think regarding the restyling and the changes I've made.
No rush, I'm not in a hurry to publish.
Thanks again again,
Rory
