### 2017-05-31 #260 @benmarwick
Thanks, I've reviewed the changes made at https://github.com/rorynolan/filesstrings/commit/dc5338deb72af5c7c3489e275efb3b6b9ec0e3bf, and they are excellent. I'm delighted to say that the author has responded to my review and made changes to my satisfaction. I recommend approving this package. Congratulations on such high quality work on a very useful package!
