### 2017-05-31 f4b9b3a
Make ExtractNumbers() return integers where appropriate.
