+++ CONDUCT.md
@@ -22,4 +22,4 @@ opening an issue or contacting one or more of the project maintainers.
 
 This Code of Conduct is adapted from the Contributor Covenant 
 (http:contributor-covenant.org), version 1.0.0, available at 
-http://contributor-covenant.org/version/1/0/0/
+http://contributor-covenant.org/version/1/0/0/.
