### 2017-05-31 5397441
Package is now peer-reviewed.
