### 2017-05-31 #260 @rorynolan
Thanks @benmarwick for the review and kind words.
Thanks @karthik for everything.
Thanks @arfon for the great JOSS service.  DOI: 10.5281/zenodo.801452
