### 2017-06-01 379e7a7
Slight modification of AllEqual documentation.
