### 2017-06-02 #260 @arfon
@rorynolan - Could you move the references you currently have in the `paper.md` file into a `paper.bib` file and cite them directly please? (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)). The citations in `paper.bib` are automatically compiled into the final PDF by Pandoc and so they don't need to be listed at the end of the `paper.md` file.
