### 2017-06-04 d362485
Rework paper to have references in bib file.
