+++ paper/paper.bib
@@ -49,7 +49,7 @@
   note = {R package version 1.5},
   url = {https://CRAN.R-project.org/package=magrittr},
 }
-@Manual{,
+@Manual{ore,
   title = {ore: An R Interface to the Oniguruma Regular Expression Library},
   author = {Jon Clayden and based on Onigmo by K. Kosako and K. Takata},
   year = {2016},
