### 2017-06-04 #260 @rorynolan
Thanks again everyone. JOSS has been a fantastic experience, I'll be recommending it to all my colleagues.
